refactor(adapters/utils): consolidate dynamic.go into shared adapters/common/utils/
The function bodies of utils/dynamic.go are byte-identical across the
three adapter versions (v0.204 / v0.215 / v0.219); the only divergence
is a per-version comment label inside `Check value types (BAML <ver>+
returns these as values instead of pointers)`. Move the file into a
shared package at adapters/common/utils/ with version-agnostic comment
text; per-adapter copies become 6-line shims that re-export
UnwrapDynamicValue so codegen-emitted callers (which qualify against
<selfPkg>/utils) keep their existing import path.

The per-adapter utils/dynamic_test.go files were already byte-identical
across all three adapters (blob 8524c3f6cb), and dynamic_bench_test.go
was byte-identical between v0.215 and v0.219 (blob 3b78935bd8). They
move into the shared package once and are deleted from each adapter.

adapters/common/go.mod gains a direct boundaryml/baml dep at v0.204.0
(the lowest-pinned version). MVS verification: each per-adapter module
still resolves its pinned BAML version via Go's minimum-version
selection (v0.204.0 / v0.215.0 / v0.219.0 unchanged).

embed.go regenerated via `go run ./cmd/embed` reflects the new shared
file (adapters/common/embed.go) and the simplified per-adapter utils
dirs (now uniformly listed as `utils` since no _test.go siblings remain
to expand into per-file embed entries).
