+++ adapters/common/embed.go
@@ -4,7 +4,7 @@ import (
 	"embed"
 )
 
-//go:embed codegen/codegen.go embed.go go.mod go.sum helpers.go
+//go:embed codegen/codegen.go embed.go go.mod go.sum helpers.go utils/dynamic.go
 var source embed.FS
 
 var Sources = make(map[string]embed.FS)
