+++ adapters/common/utils/dynamic_test.go
